requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.23.0
Flask-RESTful==0.3.7
gevent==1.4.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
Werkzeug==0.15.6
